Export functions to index.js
