+++ src/populate.js
@@ -9,9 +9,6 @@ async function createGame(){
   return id;
 }
 
-const button = document.querySelector('.submit');
-const refreshButton = document.querySelector('.refresh');
-const ul = document.querySelector('.list');
 
 async function postInfo(id, name, score) {
  const info = await axios.post(`https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/${id}/scores`, {
@@ -29,32 +26,30 @@ async function getData(id) {
 }
 
 
-createGame().then((id) => {
+// createGame().then((id) => {
   
-  button.addEventListener('click',(e) => {
-    const name = document.querySelector('.name').value;
-    const score = document.querySelector('.score').value;
-    if (name && score) {
-      e.preventDefault();
-      postInfo(id,name,score);
-      document.querySelector('form').reset();
-    }
-  });
+//   button.addEventListener('click',(e) => {
+//     const name = document.querySelector('.name').value;
+//     const score = document.querySelector('.score').value;
+//     if (name && score) {
+//       e.preventDefault();
+//       postInfo(id,name,score);
+//       document.querySelector('form').reset();
+//     }
+//   });
 
-  refreshButton.addEventListener('click', () => {
-    getData(id).then((result) => {
-      ul.innerHTML = '';
-      result.forEach((res, index) => {
-        console.log(res);
-        index % 2 === 0 ? ul.innerHTML += `<li>${res.user}:${res.score}</li>`: ul.innerHTML += `<li  class="li-grey">${res.user}:${res.score}</li>`;
+//   refreshButton.addEventListener('click', () => {
+//     getData(id).then((result) => {
+//       ul.innerHTML = '';
+//       result.forEach((res, index) => {
+//         console.log(res);
+//         index % 2 === 0 ? ul.innerHTML += `<li>${res.user}:${res.score}</li>`: ul.innerHTML += `<li  class="li-grey">${res.user}:${res.score}</li>`;
         
-      })
-    })
-  })
+//       })
+//     })
+//   })
 
 
-})
+// })
 
-
-
-
+export default {createGame, postInfo, getData};
